+++ config/rabbitmq.go
@@ -6,6 +6,7 @@ import (
 )
 
 type RabbitMQ struct {
+	Enable   bool   `mapstructure:"enable" json:"enable" yaml:"enable"`
 	Host     string `mapstructure:"host" json:"host" yaml:"host"`
 	Port     int    `mapstructure:"port" json:"port" yaml:"port"`
 	Username string `mapstructure:"username" json:"username" yaml:"username"`